+++ checkrules.js
@@ -40,22 +40,9 @@ const
       'file': 'default.js'
     },
     {
-      'name': 'ESLint Core - Next (github.com)',
-      'url': 'https://raw.githubusercontent.com/eslint/eslint/master/conf/eslint.json',
-      'docs': 'https://github.com/eslint/eslint/blob/master/docs/rules/{rule}.md',
-      'file': 'default.js'
-    },
     {
       'name': 'React Plugin - Released (unpkg.com)',
       'url': 'https://unpkg.com/eslint-plugin-react',
-      'docs': 'https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/{rule}.md',
-      'file': 'react.js',
-      'prefix': 'react/',
-      'processRemoteData': processReactPlugin
-    },
-    {
-      'name': 'React Plugin - Next (github.com)',
-      'url': 'https://raw.githubusercontent.com/yannickcr/eslint-plugin-react/master/index.js',
       'docs': 'https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/{rule}.md',
       'file': 'react.js',
       'prefix': 'react/',